+++ app/Http/Controllers/EquipmentBrandController.php
@@ -9,6 +9,7 @@ use App\Http\Resources\EquipmentBrand\ShowResource;
 use App\Models\EquipmentBrand;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class EquipmentBrandController extends Controller
 {
@@ -17,6 +18,7 @@ class EquipmentBrandController extends Controller
      */
     public function index(): AnonymousResourceCollection
     {
+        Gate::authorize('viewAny', EquipmentBrand::class);
         $equipmentBrands = EquipmentBrand::get();
         return IndexResource::collection($equipmentBrands);
     }
@@ -26,6 +28,7 @@ class EquipmentBrandController extends Controller
      */
     public function store(StoreEquipmentBrandRequest $request): ShowResource
     {
+        Gate::authorize('create', EquipmentBrand::class);
         $data = $request->validated();
         $equipmentBrand = EquipmentBrand::create($data);
         return new ShowResource($equipmentBrand);
@@ -36,6 +39,7 @@ class EquipmentBrandController extends Controller
      */
     public function show(EquipmentBrand $equipmentBrand): ShowResource
     {
+        Gate::authorize('view', $equipmentBrand);
         return new ShowResource($equipmentBrand);
     }
 
@@ -44,6 +48,7 @@ class EquipmentBrandController extends Controller
      */
     public function update(UpdateEquipmentBrandRequest $request, EquipmentBrand $equipmentBrand): ShowResource
     {
+        Gate::authorize('update', $equipmentBrand);
         $data = $request->validated();
         $equipmentBrand->update($data);
         return new ShowResource($equipmentBrand);
@@ -54,6 +59,7 @@ class EquipmentBrandController extends Controller
      */
     public function destroy(EquipmentBrand $equipmentBrand): Response
     {
+        Gate::authorize('delete', $equipmentBrand);
         $equipmentBrand->delete();
         return response()->noContent();
     }
