+++ app/Http/Controllers/EquipmentModelController.php
@@ -9,6 +9,7 @@ use App\Http\Resources\EquipmentModel\ShowResource;
 use App\Models\EquipmentModel;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class EquipmentModelController extends Controller
 {
@@ -17,6 +18,7 @@ class EquipmentModelController extends Controller
      */
     public function index(): AnonymousResourceCollection
     {
+        Gate::authorize('viewAny', EquipmentModel::class);
         $equipmentModels = EquipmentModel::get();
         return IndexResource::collection($equipmentModels);
     }
@@ -26,6 +28,7 @@ class EquipmentModelController extends Controller
      */
     public function store(StoreEquipmentModelRequest $request): ShowResource
     {
+        Gate::authorize('create', EquipmentModel::class);
         $data = $request->validated();
         $equipmentModel = EquipmentModel::create($data);
         return new ShowResource($equipmentModel);
@@ -36,6 +39,7 @@ class EquipmentModelController extends Controller
      */
     public function show(EquipmentModel $equipmentModel): ShowResource
     {
+        Gate::authorize('view', $equipmentModel);
         return new ShowResource($equipmentModel);
     }
 
@@ -44,6 +48,7 @@ class EquipmentModelController extends Controller
      */
     public function update(UpdateEquipmentModelRequest $request, EquipmentModel $equipmentModel): ShowResource
     {
+        Gate::authorize('update', $equipmentModel);
         $data = $request->validated();
         $equipmentModel->update($data);
         return new ShowResource($equipmentModel);
@@ -54,6 +59,7 @@ class EquipmentModelController extends Controller
      */
     public function destroy(EquipmentModel $equipmentModel): Response
     {
+        Gate::authorize('delete', $equipmentModel);
         $equipmentModel->delete();
         return response()->noContent();
     }
