+++ app/Http/Controllers/RoleController.php
@@ -9,6 +9,7 @@ use App\Http\Resources\Role\ShowResource;
 use App\Models\Role;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -17,6 +18,7 @@ class RoleController extends Controller
      */
     public function index(): AnonymousResourceCollection
     {
+        Gate::authorize('viewAny', Role::class);
         $roles = Role::get();
         return IndexResource::collection($roles);
     }
@@ -26,6 +28,7 @@ class RoleController extends Controller
      */
     public function store(StoreRoleRequest $request): ShowResource
     {
+        Gate::authorize('create', Role::class);
         $data = $request->validated();
         $role = Role::create($data);
         return new ShowResource($role);
@@ -36,6 +39,7 @@ class RoleController extends Controller
      */
     public function show(Role $role): ShowResource
     {
+        Gate::authorize('view', $role);
         return new ShowResource($role);
     }
 
@@ -44,6 +48,7 @@ class RoleController extends Controller
      */
     public function update(UpdateRoleRequest $request, Role $role): ShowResource
     {
+        Gate::authorize('update', $role);
         $data = $request->validated();
         $role->update($data);
         return new ShowResource($role);
@@ -54,6 +59,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role): Response
     {
+        Gate::authorize('delete', $role);
         $role->delete();
         return response()->noContent();
     }
